role.ts inside of usePermissions hook
